Fix retrieving a date stored in a DATE-typed column
Technically SQLite does not have support for other types than a very few, which do not include DATE. However, it is possible to create a DATE column, in which case anything can be put in it. Users would expect to retrieve a valid date when putting an integer as a value of such a column, especially since adding a float integral value returns an int value.